refactor: Rename index.d.ts files that shadow index.js
cf. https://github.com/microsoft/TypeScript/blob/3471bd7d867364af93e9c2f67264d6c75ec93cd5/src/compiler/commandLineParser.ts#L3701-L3709
