+++ packages/compartment-mapper/package.json
@@ -20,7 +20,7 @@
   },
   "type": "module",
   "main": "./index.js",
-  "types": "./index.d.ts",
+  "types": "./types.d.ts",
   "exports": {
     ".": "./index.js",
     "./import.js": "./import.js",
