+++ packages/daemon/types.d.ts
