+++ packages/lp32/package.json
@@ -27,7 +27,7 @@
   "type": "module",
   "main": "./index.js",
   "module": "./index.js",
-  "types": "./index.d.ts",
+  "types": "./types.d.ts",
   "exports": {
     ".": "./index.js",
     "./reader.js": "./reader.js",
