+++ packages/static-module-record/types.d.ts
