+++ packages/stream/package.json
@@ -23,7 +23,7 @@
   "type": "module",
   "main": "./index.js",
   "module": "./index.js",
-  "types": "./index.d.ts",
+  "types": "./types.d.ts",
   "exports": {
     ".": "./index.js",
     "./package.json": "./package.json"
