--- packages/where/index.d.ts
@@ -1,6 +1,0 @@
-export {
-  whereEndoState,
-  whereEndoEphemeralState,
-  whereEndoSock,
-  whereEndoCache,
-} from './types.js';
